[BUGFIX] load `sub/index.md` if `sub.md` not found
According to the documentation, when going to url `/sub` the core looks first for
`content/sub.md` and then for `content/sub/index.md` if the first one doesn't
exist. That didn't work before, but that's now fixed.

There was actually a leading slash `/` missing when appending `index` to the
path name.

Resolves: #126
Documentation: #126
Releases: 1.3.0
Related: #126, #117, #119