Delete 'cache' and 'datastorage' directories from repository and only create them when required
This addresses the issue mentioned in https://github.com/PhileCMS/Phile/pull/182: that creating a 'datastorage' directory on the fly does not recreate everything that was in there in the original repository – notably, a .htaccess file.

The way I've set it up now is to remove the 'cache' and 'datastorage' directories from the repository altogether, and to only create them when we need them.

I've addressed the security issue by preventing access to the whole 'lib/' directory, and to the whole 'plugins/' directory while we're at it. I don't think that there's anything in there that needs to be accessible from the outside (right?).