Add gitattributes
With `.gitattributes` it's now easy to package the repository. The problem
with installing it through composer is that it includes the entire contents of
the git branch, including potentially undesirable files such as `tests/` and
`.travis.yml`. I've therefore added a `.gitattributes` which will ignore
this kind of files when creating a new Phile-project through composer.

These files are also ignored when zipping the repository.