Moves call to $app->bootstrap out of $server->run
Serving the request should happen on a ready Phile-app. The issue came apparent
when test cases did run bootstrap twice.

For testing plugins have to be loaded early so their facility (classes,
function, ...) are available in the test-cases: bootstrapping them. Serving a request in test-cases
did run bootstrap - and with that plugin-loading - a second time which led to
erroneous side effects in the global configuration (populating plugin settings).