fixes Pages without meta block are rendered empty #351
The raw-content has to be split into meta and content.

This was done in two places before (Page and Parser) essentially
duplicating the same functionality. Now this responsibility is delegated to the
Parser, it is the party with the knowledge on how to detect and extract the
meta from the content.

The evaluation of meta and content is still done twice as before, so there's
still room for performance improvement. But at least the implementations are
merged into one code place.